Support arbitrary exception payload for calling Dart methods from C++ (#341)
Updated Dart and FFI templates to return error value through an "out"
parameter when calling a throwing Dart method from C++. Previously the
error value was returned through a return value, which is restricted to
an integer type due to Dart's own requirements for "call Dart from C++"
system.

New implementation uses the type-restricted return value only to carry a
Boolean "error happened" flag. Returning error value through an "out"
parameter allows for arbitrary types as exception payload, similar to
how it already works for Java and Swift.

Added functional tests. Updated smoke tests.

Resolves: #140

+++ gluecodium/src/test/resources/smoke/comments/output/dart/lib/src/smoke/DeprecationCommentsOnly.dart
@@ -257,9 +257,9 @@ Pointer<Void> smoke_DeprecationCommentsOnly_toFfi(DeprecationCommentsOnly value)
     __lib.cacheObject(value),
     __lib.LibraryContext.isolateId,
     __lib.uncacheObjectFfi,
-    Pointer.fromFunction<Int64 Function(Uint64, Pointer<Void>, Pointer<Uint8>)>(_DeprecationCommentsOnly_someMethodWithAllComments_static, __lib.unknownError),
-    Pointer.fromFunction<Int64 Function(Uint64, Pointer<Uint8>)>(_DeprecationCommentsOnly_isSomeProperty_get_static, __lib.unknownError),
-    Pointer.fromFunction<Int64 Function(Uint64, Uint8)>(_DeprecationCommentsOnly_isSomeProperty_set_static, __lib.unknownError)
+    Pointer.fromFunction<Uint8 Function(Uint64, Pointer<Void>, Pointer<Uint8>)>(_DeprecationCommentsOnly_someMethodWithAllComments_static, __lib.unknownError),
+    Pointer.fromFunction<Uint8 Function(Uint64, Pointer<Uint8>)>(_DeprecationCommentsOnly_isSomeProperty_get_static, __lib.unknownError),
+    Pointer.fromFunction<Uint8 Function(Uint64, Uint8)>(_DeprecationCommentsOnly_isSomeProperty_set_static, __lib.unknownError)
   );
   __lib.reverseCache[_smoke_DeprecationCommentsOnly_get_raw_pointer(result)] = value;
   return result;
